requirements.txt
aiogram==2.9.2
python-dotenv==0.21.1
pydantic==1.6.1
pillow==10.0.0